team/requirements.txt
fastapi
uvicorn
python-arango
pydantic
shared_models @ file:///shared_models
